added wrappers around builders to provide same API
